# openapi.yaml
openapi: 3.0.0
info:
  title: TradeOn API
  version: 1.0.0
  description: 'TradeOn은 사용자 간 물물교환 기반의 중고 거래 플랫폼입니다.


    이 API 문서는 다음과 같은 기능들을 포함합니다:

    - 사용자 회원가입, 로그인 및 정보 조회

    - 교환 요청 및 거래 이력 관리

    - 사용자 리뷰 시스템

    - 가치 평가 게시판

    - 하이브리드 검색기능

    - 게시물 등록/삭제/조회

    - 채팅 및 거래 약속

    - 거래 데이터 분석 및 시각화


    JWT 인증이 필요한 API의 경우 Authorize 버튼을 통해 토큰을 입력해주세요.'
servers:
- url: https://673e-121-140-48-141.ngrok-free.app
  description: 로컬
...
paths:
...
  /api/v1/valuations:
    get:
      tags:
        - Valuation
      summary: 가치 평가 게시글 목록 조회
      description: 인증된 사용자의 가치 평가 게시글을 페이지네이션하여 최신순으로 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 페이지 번호 (기본값 1)
          required: false
          schema:
            type: integer
        - name: per_page
          in: query
          description: 페이지당 게시글 수 (기본값 10)
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 게시글 목록 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValuationPost'
        '401':
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Valuation
      summary: 가치 평가 게시글 작성
      description: 인증된 사용자가 새 가치 평가 게시글을 등록합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValuationPostCreateRequest'
      responses:
        '201':
          description: 게시글 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationPost'
        '400':
          description: 필수 입력 항목 누락 또는 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/valuations/{postid}:
    get:
      tags:
        - Valuation
      summary: 가치 평가 게시글 상세 조회
      description: 특정 게시글의 상세 정보를 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 조회할 게시글 ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 게시글 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationPost'
        '401':
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Valuation
      summary: 가치 평가 게시글 삭제
      description: 작성자만 자신의 가치 평가 게시글을 삭제할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 삭제할 게시글 ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 게시글 삭제 성공
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/valuations/{postid}/price:
    post:
      tags:
        - Valuation
      summary: 평가 의견 등록
      description: 특정 게시글에 대한 가격 의견을 등록합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 평가할 게시글 ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValuationOpinionCreateRequest'
      responses:
        '201':
          description: 의견 등록 성공
        '400':
          description: 필수 입력 항목 누락 또는 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/valuations/{postid}/average:
    get:
      tags:
        - Valuation
      summary: 게시글의 평균 가격 조회
      description: 특정 가치 평가 게시글에 대한 유효한 의견들의 평균 가격을 반환합니다. JWT 인증 필요.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          required: true
          description: 평균 가격을 조회할 게시글 ID
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: 평균 가격 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationAverageResponse'
        '404':
          description: 게시글을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ValuationPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        owner_id:
          type: integer
        created_at:
          type: string
          format: date-time
    ValuationPostCreateRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          example: 이 기타의 가치는 어느 정도일까요?
        description:
          type: string
          example: Fender 기타, 거의 새 제품입니다.
    ValuationOpinionCreateRequest:
      type: object
      required:
        - price
      properties:
        price:
          type: integer
          example: 600000
    ValuationAverageResponse:
      type: object
      properties:
        postId:
          type: integer
          example: 123
        validCount:
          type: integer
          example: 5
        averagePrice:
          type: number
          format: float
          example: 28700.0
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        msg:
          type: string
          example: 오류 메시지입니다.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []